Added a new Route duration test. Fixed a bug with implicit casting in DefaultReport.java where the average delivery time would always be an int.

+++ src/main/java/nz/ac/victoria/ecs/kpsmart/reporting/impl/DefaultReport.java
@@ -101,7 +101,7 @@ public class DefaultReport implements Report {
 							continue;
 						
 						revinue += price.getCost(m);
-						totalDeliveryTime += m.getShippingDuration()/msinhour;
+						totalDeliveryTime += m.getShippingDuration()/(double)msinhour;
 						mailCount++;
 						
 						for (Route r : m.getRoute())
